+++ fabric/src/main/java/io/github/sakurawald/module/mixin/gameplay/multi_obsidian_platform/EntityMixin.java
@@ -1,6 +1,9 @@
 package io.github.sakurawald.module.mixin.gameplay.multi_obsidian_platform;
 
 #if MC_VER <= MC_1_20_6
+import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import io.github.sakurawald.module.initializer.gameplay.multi_obsidian_platform.MultiObsidianPlatformInitializer;
 import net.minecraft.block.Blocks;
 import net.minecraft.entity.Entity;
@@ -10,7 +13,6 @@ import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
 
 @Mixin(Entity.class)
 public abstract class EntityMixin {
@@ -27,8 +29,8 @@ public abstract class EntityMixin {
         return entity.getWorld();
     }
 
-    @Redirect(method = "getTeleportTarget", at = @At(value = "FIELD", target = "Lnet/minecraft/server/world/ServerWorld;END_SPAWN_POS:Lnet/minecraft/util/math/BlockPos;"), require = 1)
-    BlockPos $findDimensionEntryPoint(ServerWorld toLevel) {
+    @ModifyExpressionValue(method = "getTeleportTarget", at = @At(value = "FIELD", target = "Lnet/minecraft/server/world/ServerWorld;END_SPAWN_POS:Lnet/minecraft/util/math/BlockPos;"), require = 1)
+    BlockPos $findDimensionEntryPoint(BlockPos original) {
         // modify: resource_world:overworld -> minecraft:the_end (default obsidian platform)
         // feature: https://bugs.mojang.com/browse/MC-252361
         if (getEntityCurrentLevel().getRegistryKey() != World.OVERWORLD) return ServerWorld.END_SPAWN_POS;
@@ -44,9 +46,15 @@ public abstract class EntityMixin {
         BlockPos.iterate(i - 2, j, k - 2, i + 2, j, k + 2).forEach(blockPos -> serverLevel.setBlockState(blockPos, Blocks.OBSIDIAN.getDefaultState()));
     }
 
-    /* This method will NOT be called when a PLAYER jump into overworld's ender-portal-frame */
-    @Redirect(method = "moveToWorld", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/world/ServerWorld;createEndSpawnPlatform(Lnet/minecraft/server/world/ServerWorld;)V"), require = 1)
-    public void $changeDimension(ServerWorld toLevel) {
+    /*
+    1. In vanilla Minecraft, when player entity and non-player entity jump into the Ender Portal Frame, the obsidian platform will be re-created.
+    2. For ServerPlayerEntity, the obsidian platform location is the player's teleport location.
+    3. For Entity, the obsidian platform location is the fixed location END_SPAWN_POS.
+
+    This method will NOT be called when a PLAYER jump into overworld's ender-portal-frame */
+    @WrapOperation(method = "moveToWorld", at = @At(value = "INVOKE", target = "Lnet/minecraft/server/world/ServerWorld;createEndSpawnPlatform(Lnet/minecraft/server/world/ServerWorld;)V"), require = 0)
+    // NOTE: require = 0, due to mixin fails in forge platform.
+    private void makeObsidianPlatformWhenNonPlayerEntityPassThroughThePortal(ServerWorld toLevel, Operation<Void> original) {
         // modify: resource_world:overworld -> minecraft:the_end (default obsidian platform)
         if (getEntityCurrentLevel().getRegistryKey() != World.OVERWORLD) {
             ServerWorld.createEndSpawnPlatform(toLevel);
@@ -54,6 +62,7 @@ public abstract class EntityMixin {
         }
         makeObsidianPlatform(toLevel, getTransformedEndSpawnPoint());
     }
+
 }
 #elif MC_VER > MC_1_20_6