+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - (<= MC 1.21) [nametag] fix: should not block the player using `nether portal` or `the end portal`.
+- [gameplay.multi_obsidian_platform] feature: make it working on `forge` platform.
+- [command_toolbox.hat] feature: make it working on `forge` platform.